Moved model_validation tests to invalid_models_tests.
